fix: removed non-required abstraction
